rag_agent/src/requirements.txt
notion-client==2.2.1
openai==1.12.0
tiktoken==0.5.2
nltk==3.8.1
azure-search-documents==11.4.0
azure-identity==1.15.0
python-dotenv==1.0.0
pandas
tenacity
chromadb==0.5.15
pillow==10.1.0
promptflow==1.8.0